# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API documentation for the News Aggregator project",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "description": "User login and return JWT token",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "JWT token",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "500": {
                        "description": "Could not create token"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user details",
                "description": "Get authenticated user details",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "User details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "description": "Update user details",
                "operationId": "updateUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User registration",
                "description": "Register a new user",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "title": "Login Request",
                "description": "Request parameters for login",
                "properties": {
                    "email": {
                        "description": "Email of user",
                        "type": "string"
                    },
                    "password": {
                        "description": "Password of the user",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "Register Request",
                "description": "Request parameters for user registration",
                "properties": {
                    "name": {
                        "description": "Name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email of the user",
                        "type": "string"
                    },
                    "password": {
                        "description": "Password of the user",
                        "type": "string"
                    },
                    "authors": {
                        "description": "Array of author UUIDs",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    "categories": {
                        "description": "Array of category UUIDs",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    "sources": {
                        "description": "Array of source UUIDs",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdateUserRequest": {
                "title": "Update User Request",
                "description": "Request parameters for updating user data",
                "properties": {
                    "name": {
                        "description": "Name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email of the user",
                        "type": "string"
                    },
                    "password": {
                        "description": "New password for the user",
                        "type": "string"
                    },
                    "authors": {
                        "description": "Array of author UUIDs",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    "categories": {
                        "description": "Array of category UUIDs",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    "sources": {
                        "description": "Array of source UUIDs",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                },
                "type": "object"
            },
...
            "Author": {
                "title": "Author",
                "description": "Author model representing an author",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "UUID of the author",
                        "type": "string"
                    },
                    "name": {
                        "description": "Author name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "description": "Category model representing a category",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "UUID of the category",
                        "type": "string"
                    },
                    "name": {
                        "description": "Category name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Source": {
                "title": "Source",
                "description": "Source model representing a source",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "UUID of the source",
                        "type": "string"
                    },
                    "external_id": {
                        "description": "Source external id",
                        "type": "string"
                    },
                    "name": {
                        "description": "Source name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model representing a registered user",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "UUID of the user",
                        "type": "string"
                    },
                    "name": {
                        "description": "Name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email of the user",
                        "type": "string"
                    },
                    "preferredAuthors": {
                        "description": "Array of preferred authors",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Author"
                        }
                    },
                    "preferredCategories": {
                        "description": "Array of preferred categories",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "preferredSources": {
                        "description": "Array of preferred sources",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Source"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}